Move GraphQL item class and cleanup cache test
